// papertrlui-v2/src/app/shared/enums/app-action-label.ts
export enum AppActionLabel {
  ACTION_LABEL_ATTACH_TO_DOCUMENT = 'Manage Documents',
  ACTION_LABEL_DELETE = 'Delete',
  ACTION_LABEL_RE_SYNC = 'Re-Sync',
  ACTION_LABEL_REJECT = 'Reject',
  ACTION_LABEL_GENERATE_BILL = 'Generate Bill',
  ACTION_LABEL_LOCK = 'Lock',
  ACTION_LABEL_UNLOCK = 'Unlock',
  ACTION_LABEL_APPROVE = 'Approve',
  ACTION_LABEL_CREATE_INVOICE = 'Create an Invoice',
  ACTION_LABEL_APPROVE_REJECT = 'Approve or Reject',
  ACTION_LABEL_SKIP_APPROVAL = 'Skip Approval',
  ACTION_LABEL_DETAIL_VIEW = 'Detail View',
  ACTION_LABEL_VIEW_COMMENTS = 'View Comments',
  ACTION_LABEL_APPLY_TO_BILL = 'Apply to Bill',
  ACTION_LABEL_VIEW_APPLIED_BILL = 'View Bills',
  ACTION_LABEL_SEND_TO_CUSTOMER = 'Send to Customer',
  ACTION_LABEL_VIEW_CREDIT_NOTE = 'View Credit Notes',
  ACTION_LABEL_CREATE_CREDIT_NOTE = 'Create Credit Note',
  ACTION_APPLY_CREDIT_NOTE = 'Apply Credit Note',
  PAYMENT_PROCESS = 'Mark As Offline Payment Processing',
  PAYMENT_UNPROCESS = 'Undo Offline Payment Processing',
  ACTION_LABEL_CLONE_ROLE = 'Clone Role',
  ACTION_LABEL_USE = 'Use',
  ACTION_LABEL_PAY_NOW = 'Pay Now',
  ACTION_LABEL_AUDIT_TRAIL = 'Audit Trail',
  ACTION_LABEL_SPLIT_TRANSACTION = 'Split Transaction',
  ACTION_LABEL_ACTIVITY_LOG = 'Activity Log',
  ACTION_LABEL_CANCEL_PAYMENT = 'Cancel Payment',
  ACTION_LABEL_CANCEL = 'Cancel',
  ACTION_LABEL_GENERATE_RECEIPT = 'Generate PO receipt',
  ACTION_LABEL_DOWNLOAD_ATTACHMENT = 'Download Receipt',
  ACTION_LABEL_VIEW_SUMMARY = 'View Summary',
  ACTION_LABEL_DOWNLOAD_BILL = 'Download Bill',
  ACTION_LABEL_DOWNLOAD_INVOICE = 'Download Invoice',
  ACTION_LABEL_DOWNLOAD_RECEIPT = 'Download Receipt',
  ACTION_LABEL_UNDO_APPROVAL = 'Undo',
  ACTION_GENERATE_DETAILED_REPORT = 'Download Detailed Report',
  ACTION_LABEL_RESEND_INVITATION = 'Resend Invitation',
  ACTION_LABEL_SUBMIT_FOR_APPROVAL = 'Submit for Approval',
  ACTION_LABEL_RESEND_EMAIL = 'Resend Email',
  ACTION_LABEL_UNDO_REJECTION = 'Undo Rejection',
  ACTION_LABEL_VIEW_REPORT = 'Download Report',
  SEND_PO_ATTACHMENT = 'Email to Vendor',
  ACTION_LABEL_EDIT_RESUBMIT = 'Edit & Resubmit',
  ACTION_LABEL_CLOSE = 'Close',
  ACTION_LABEL_RE_OPEN = 'Reopen',
  ACTION_LABEL_MARK_AS_MAILED = 'Mark As Mailed',
  ACTION_LABEL_VOID_PAYMENT = 'Cancel Transaction',
  ACTION_LABEL_CHANGE_BILL = 'Change Bill',
  ACTION_LABEL_CHANGE_DOCUMENT = 'Change Document',
  ACTION_LABEL_VIEW_BILL = 'View Bill',
  ACTION_LABEL_VIEW_ATTACHMENT = 'View Receipt',
  ACTION_LABEL_ADD_NOTE = 'Add Note',
  ACTION_LABEL_EDIT = 'Edit',
  ACTION_LABEL_SEND_ACH_INFORMATION_REQUEST = 'Send ACH Info Request ',
  ACTION_LABEL_TOP_UP = 'Top Up',
  ACTION_LABEL_DISCARD = 'Discard',
  ACTION_LABEL_CHANGE_EMPLOYEE = 'Change Employee',
  ACTION_LABEL_PROCESS_STATEMENT = 'Process Transaction',
  ACTION_LABEL_CHANGE_ASSIGNEE = 'Change Assignee',
  ACTION_LABEL_ACTIVATE = 'Activate',
  ACTION_LABEL_FREEZE = 'Freeze',
  ACTION_LABEL_INACTIVATE = 'Inactivate',
  ACTION_LABEL_PASSWORD_RESET = 'Password Reset',
  ACTION_SEND_INVITATION = 'Send Vendor Invitation',
  ACTION_IMPORT_VENDORS = 'Import Vendors',
  ACTION_EXPORT_VENDORS = 'Export Vendors',
  ACTION_VENDOR_REQUEST = 'Vendor Requests',
  ACTION_COMMUNITY_VENDOR = 'Community Vendors',
  ACTION_VENDOR_GROUPS = 'Vendor Groups',
  ACTION_LABEL_APPLIED_PAYMENTS = 'Apply Payments',
  ACTION_ATTACHED_TO_BILL = 'Attach To a Bill',
  ACTION_CREATE_BILL = 'Create a Bill',
  ACTION_CREATE_DIGITAL_CARD = 'Create a Digital Card',
  ACTION_CREATE_INVOICE = 'Create an Invoice',
  ACTION_ADD_TO_LOCAL = 'Add To Local',
  ACTION_LABEL_SEND_TO_VENDOR_APPROVAL = 'Send to Vendor Approval',
  ACTION_LABEL_EXPORT= 'Export',
  ACTION_LABEL_EDIT_TENANT_MENU_PRIVILEGES= 'Edit Tenant Menu Privileges',
  ACTION_LABEL_USERS_ACTIVITY_LOGS= 'Users Activity Logs',
  ACTION_DOWNLOAD_SELECTED = 'Download Selected',
  ACTION_DOWNLOAD_All = 'Download All',
  ACTION_EXPORT_SELECTED = 'Export Selected',
  ACTION_EXPORT_All = 'Export All',
  TO_PROCESS_ATTACHED_TO = 'Attach to',
  TO_PROCESS_DELETE = 'Delete',
  DELETED_LIST_DELETE = 'Delete',
  DELETED_LIST_RECOVER = 'Recover',
  ACTION_LABEL_MARK_AS_PAID = 'Mark as Paid',
  ACTION_LABEL_MARK_AS_UNPAID = 'Mark as Unpaid',
  ACTION_LABEL_MARK_AS_DEFAULT = 'Mark as Default'
}