// papertrlui-v2/src/app/shared/enums/app-bulk-button.ts
export enum AppBulkButton {
  BUTTON_ACTIVATE = 'ACTIVATE',
  BUTTON_INACTIVATE = `INACTIVATE`,
  BUTTON_DELETE = 'DELETE',
  BUTTON_REJECT = 'REJECT',
  RESEND_REQUEST = 'RESEND_REQUEST',
  BUTTON_EXPORT = 'EXPORT',
  BUTTON_DOWNLOAD = 'DOWNLOAD',
  BUTTON_QUICK_APPROVE = 'QUICK_APPROVE',
  APPLY_CREDIT_NOTE = 'APPLY_CREDIT_NOTE',
  PAYMENT_PROCESS = 'PAYMENT_PROCESS',
  BUTTON_APPROVE = 'APPROVE',
  BUTTON_SEND_INVITATION = 'SEND_INVITATION',
  BUTTON_ADD = 'ADD',
  BUTTON_SYNC = 'SYNC',
  BUTTON_OPEN = 'OPEN',
  BUTTON_CLOSE = 'CLOSE',
  PAY_NOW = 'PAY_NOW',
  SUBMIT_FOR_APPROVAL = 'SUBMIT_FOR_APPROVAL',
  MAKE_AN_OFFLINE_PAYMENT = 'MAKE_AN_OFFLINE_PAYMENT',
  INBOX_SAVE_AND_CONTINUE = 'INBOX_SAVE_AND_CONTINUE',
  INBOX_SUBMIT_AND_CONTINUE = 'INBOX_SUBMIT_AND_CONTINUE',
  INBOX_SAVE_AS_APPROVED = 'SAVE_AS_APPROVED_FROM_INBOX',
  INBOX_SUBMIT = 'SUBMIT_FOR_APPROVED_FROM_INBOX',
  INBOX_RESET = 'RESET',
  SEND_TO_VENDOR = 'SEND_TO_VENDOR',
  SEND_PO_ATTACHMENT = 'SEND_PO_ATTACHMENT',
  BUTTON_DOWNLOAD_DETAIL_REPORT = 'BUTTON_DOWNLOAD_DETAIL_REPORT',
  BUTTON_PROCESS = 'BUTTON_PROCESS',





}